Minor tweak for better Win32 compatibility
git-svn-id: https://vmmlib.svn.sourceforge.net/svnroot/vmmlib/trunk@128 4d770e62-bb1e-0410-ad62-f2382d3e8790

+++ include/vmmlib/stringUtils.h
@@ -18,7 +18,7 @@ namespace vmml
 
 namespace stringUtils
 {
-inline ssize_t toInt( const std::string& source )
+inline long toInt( const std::string& source )
 {
     return atol( source.c_str() );
 };
